fix file sarai Userstory1Test.java and rename Userstory1.java
